Add support for lbcd & much more

+++ cmd/block.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"database/sql"
 	"fmt"
 	"strconv"
 	"time"
@@ -12,7 +13,7 @@ import (
 	"github.com/lbryio/chainquery/model"
 	"github.com/lbryio/chainquery/util"
 
-	"github.com/lbryio/lbry.go/extras/errors"
+	"github.com/lbryio/lbry.go/v2/extras/errors"
 
 	"github.com/pkg/profile"
 	"github.com/sirupsen/logrus"
@@ -35,13 +36,13 @@ var blockCmd = &cobra.Command{
 		}
 		blockHeight, err := strconv.ParseInt(args[0], 10, 64)
 		if err != nil {
-			logrus.Panic(errors.Prefix("Could not parse block height passed: ", err))
+			logrus.Panic(errors.Prefix("Could not parse block height passed", err))
 		}
 		lbrycrdClient := lbrycrd.Init()
 		defer lbrycrdClient.Shutdown()
 		blockHash, err := lbrycrdClient.GetBlockHash(blockHeight)
 		if err != nil {
-			logrus.Panic(errors.Prefix(fmt.Sprintf("Could not get block hash @ height %d: ", blockHeight), err))
+			logrus.Panic(errors.Prefix(fmt.Sprintf("Could not get block hash @ height %d", blockHeight), err))
 		}
 		//Main Chainquery DB connection
 		dbInstance, err := db.Init(config.GetMySQLDSN(), config.GetDebugQueryMode())
@@ -51,14 +52,16 @@ var blockCmd = &cobra.Command{
 		defer db.CloseDB(dbInstance)
 		logrus.Infof("Running processor on block %d with hash %s", blockHeight, blockHash)
 		block, err := model.Blocks(model.BlockWhere.Hash.EQ(blockHash.String())).OneG()
-		if err != nil {
+		if err != nil && !errors.Is(err, sql.ErrNoRows) {
 			logrus.Panic(errors.Err(err))
 		}
-		err = block.DeleteG()
-		if err != nil {
-			logrus.Fatal(errors.Err(err))
+		if block != nil {
+			err = block.DeleteG()
+			if err != nil {
+				logrus.Fatal(errors.Err(err))
+			}
+			logrus.Info("Block successfully removed")
 		}
-		logrus.Info("Block successfully removed")
 		jsonBlock, err := lbrycrd.GetBlock(blockHash.String())
 		if err != nil {
 			logrus.Fatal(errors.Err(err))
